perfect_foresight_problem MEX: forward exceptions thrown when evaluating the model .m files
This is useful when there are numerical errors in some files, such as
evaluating erf() on a complex number (under MATLAB this throws an exception;
under Octave, it instead returns a complex number).

Forward both the error message and error ID (but not other properties of the
MException object, there does not seem to be a way to do that from a MEX file).