Manual: fix method signatures for in-place dates/dseries modifying methods
Commit d1dc66611f4a9354b3af4f0a379509cc3e0a87fd incorrectly removed the first
argument on which these methods operate. That commit should only have fixed the
signatures of dates methods append_(), pop_(), remove_(), sort_() and
unique_(), which were featuring a misleading output argument (because in most
cases we don’t need that output, since it is equal to the first input argument
after modification).