Perfect foresight: harmonize the maximum error returned by LBJ
LBJ would store in oo_ the absolute error on the change of variables (on which
is based its convergence criterion), instead of the maximum error of residuals
as do the other algorithms. This commit harmonizes the output by having LBJ
store the maximum error of residuals (while not changing its convergence
criterion).