perfect_foresight_block_problem MEX: fix illegal memory access under Octave
Under Octave, when we call the block.dynamic_NN.m file, y (resp. T) in output
uses the same internal data array as y (resp. T) in input, if the array has not
been modified by the function. Hence, before asking for the destruction of the
input array, we must check that it does not share its data with the output
array.

Closes: #1996