k-order MEX: rewrite parallelization engine using a thread pool
The new parallelization engine spawns a thread pool upon startup, and use these
for the whole duration of the program, instead of spawning new threads for every
job group as was doing the previous engine.

This makes a significant performance under MATLAB for Windows, where spawning
threads in a MEX is costly.

Closes: #1962