Testsuite: in the optimizer tests, do not count function calls for algorithms provided by MATLAB
The direct caller of our function is not documented and can change across
MATLAB versions (and for 1 and 3, it has changed between R2020a and R2025b). In
any case, we do not have to test that MATLAB optimizers correctly counts
function calls, that’s not the purpose of our testsuite.