Testsuite: increase tolerance for run_block_bytecode_tests
Since we’re comparing two paths, it makes sense to use twice Δx as the
tolerance, since they each might diverge from the true path from Δx.

This is actually needed to successfully pass the test with PARDISO on block
decomposition (which, by the way, does not use Δx as the convergence criterion,
but Δf).