🐛 libkordersim: fix illegal memory read
The code was apparently assuming short-circuiting behaviour of the .and.
operator to avoid reading beyond the array boundaries. But short-circuiting is
not guaranteed in Fortran, so we need to use two different conditional
constructs.

By the way, the “do while” statement is considered as deprecated by Metcalf and
al. (2023), even though it is still part of the standard. So rather use a
combination of “do” and “if”.